Combine request and response stream base type
