+++ src/Middleware/HttpLogging/src/BufferingStream.cs
@@ -1,0 +1,202 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using System.Buffers;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.IO;
+using System.IO.Pipelines;
+using System.Linq;
+using System.Runtime.CompilerServices;
+using System.Text;
+using System.Threading.Tasks;
+using Microsoft.Extensions.Logging;
+
+namespace Microsoft.AspNetCore.HttpLogging
+{
+    internal class BufferingStream : Stream, IBufferWriter<byte>
+    {
+        private const int MinimumBufferSize = 4096; // 4K
+        private readonly ILogger _logger;
+        private int _bytesWritten;
+        private BufferSegment? _head;
+        private BufferSegment? _tail;
+        protected Memory<byte> _tailMemory; // remainder of tail memory
+        protected int _tailBytesBuffered;
+
+        public BufferingStream(ILogger logger)
+        {
+            _logger = logger;
+        }
+
+        public override bool CanRead => throw new NotImplementedException();
+
+        public override bool CanSeek => throw new NotImplementedException();
+
+        public override bool CanWrite => throw new NotImplementedException();
+
+        public override long Length => throw new NotImplementedException();
+
+        public override long Position { get => throw new NotImplementedException(); set => throw new NotImplementedException(); }
+
+        public void LogString(Encoding? encoding)
+        {
+            if (_head == null || _tail == null || encoding == null)
+            {
+                return;
+            }
+
+            // Only place where we are actually using the buffered data.
+            // update tail here.
+            _tail.End = _tailBytesBuffered;
+
+            var ros = new ReadOnlySequence<byte>(_head, 0, _tail, _tailBytesBuffered);
+
+            // If encoding is nullable
+            var body = encoding.GetString(ros);
+            _logger.LogInformation(LoggerEventIds.RequestBody, CoreStrings.RequestBody, body);
+
+            Reset();
+        }
+
+        public void Advance(int bytes)
+        {
+            if ((uint)bytes > (uint)_tailMemory.Length)
+            {
+                ThrowArgumentOutOfRangeException(nameof(bytes));
+            }
+
+            _tailBytesBuffered += bytes;
+            _bytesWritten += bytes;
+            _tailMemory = _tailMemory.Slice(bytes);
+        }
+
+        public Memory<byte> GetMemory(int sizeHint = 0)
+        {
+            AllocateMemoryUnsynchronized(sizeHint);
+            return _tailMemory;
+        }
+
+        public Span<byte> GetSpan(int sizeHint = 0)
+        {
+            AllocateMemoryUnsynchronized(sizeHint);
+            return _tailMemory.Span;
+        }
+        private void AllocateMemoryUnsynchronized(int sizeHint)
+        {
+            if (_head == null)
+            {
+                // We need to allocate memory to write since nobody has written before
+                BufferSegment newSegment = AllocateSegmentUnsynchronized(sizeHint);
+
+                // Set all the pointers
+                _head = _tail = newSegment;
+                _tailBytesBuffered = 0;
+            }
+            else
+            {
+                int bytesLeftInBuffer = _tailMemory.Length;
+
+                if (bytesLeftInBuffer == 0 || bytesLeftInBuffer < sizeHint)
+                {
+                    Debug.Assert(_tail != null);
+
+                    if (_tailBytesBuffered > 0)
+                    {
+                        // Flush buffered data to the segment
+                        _tail.End += _tailBytesBuffered;
+                        _tailBytesBuffered = 0;
+                    }
+
+                    BufferSegment newSegment = AllocateSegmentUnsynchronized(sizeHint);
+
+                    _tail.SetNext(newSegment);
+                    _tail = newSegment;
+                }
+            }
+        }
+
+        private BufferSegment AllocateSegmentUnsynchronized(int sizeHint)
+        {
+            BufferSegment newSegment = CreateSegmentUnsynchronized();
+
+            // We can't use the recommended pool so use the ArrayPool
+            newSegment.SetOwnedMemory(ArrayPool<byte>.Shared.Rent(GetSegmentSize(sizeHint)));
+
+            _tailMemory = newSegment.AvailableMemory;
+
+            return newSegment;
+        }
+
+        private BufferSegment CreateSegmentUnsynchronized()
+        {
+            return new BufferSegment();
+        }
+
+        private static int GetSegmentSize(int sizeHint, int maxBufferSize = int.MaxValue)
+        {
+            // First we need to handle case where hint is smaller than minimum segment size
+            sizeHint = Math.Max(MinimumBufferSize, sizeHint);
+            // After that adjust it to fit into pools max buffer size
+            var adjustedToMaximumSize = Math.Min(maxBufferSize, sizeHint);
+            return adjustedToMaximumSize;
+        }
+
+        protected override void Dispose(bool disposing)
+        {
+            if (disposing)
+            {
+                Reset();
+            }
+        }
+
+        public void Reset()
+        {
+            var segment = _head;
+            while (segment != null)
+            {
+                var returnSegment = segment;
+                segment = segment.NextSegment;
+
+                // We haven't reached the tail of the linked list yet, so we can always return the returnSegment.
+                returnSegment.ResetMemory();
+            }
+
+            _head = _tail = null;
+
+            _bytesWritten = 0;
+            _tailBytesBuffered = 0;
+        }
+
+        // Copied from https://github.com/dotnet/corefx/blob/de3902bb56f1254ec1af4bf7d092fc2c048734cc/src/System.Memory/src/System/ThrowHelper.cs
+        private static void ThrowArgumentOutOfRangeException(string argumentName) { throw CreateArgumentOutOfRangeException(argumentName); }
+        [MethodImpl(MethodImplOptions.NoInlining)]
+        private static Exception CreateArgumentOutOfRangeException(string argumentName) { return new ArgumentOutOfRangeException(argumentName); }
+
+        public override void Flush()
+        {
+            throw new NotImplementedException();
+        }
+
+        public override int Read(byte[] buffer, int offset, int count)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override long Seek(long offset, SeekOrigin origin)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override void SetLength(long value)
+        {
+            throw new NotImplementedException();
+        }
+
+        public override void Write(byte[] buffer, int offset, int count)
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
